+++ tests/testthat/test_compareCols.R
@@ -1,7 +1,7 @@
 ## library(devtools)
 ## setwd("tests/testthat")
 ## load_all()
-
+library(data.table)
 context("compareCols")
 
 test_that("basic",{
@@ -9,8 +9,9 @@ test_that("basic",{
     fileRef <- "testReference/compareCols_1.rds"
 
     pk <- readRDS(file=system.file("examples/data/xgxr2.rds",package="NMdata"))
-    pk.reduced <- copy(pk)
-    pk.reduced <- pk.reduced[1:(.N%/%2)]
+    ## pk.reduced <- copy(pk)
+    ## pk.reduced <- pk.reduced[1:(.N%/%2)]
+    pk.reduced <- pk[1:(.N%/%2)]
     pk.reduced[,CYCLE:=NULL]
     pk.reduced[,AMT:=as.character(AMT)]